requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
docopt==0.6.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pipreqs==0.4.10
python-dateutil==2.8.1
requests==2.25.1
scapy==2.4.3
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
yarg==0.1.9
